# FundooNotes/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "rushipatil6632@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/addCollaboratorByNoteId": {
            "post": {
                "summary": "Add Colaborator ",
                "description": " Add Colaborator a to specific Note ",
                "operationId": "3ca0b9d0e282a1181aa23151f6bffda9",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "note_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "note_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collaborator Created Sucessfully"
                    },
                    "202": {
                        "description": "Collab Not Added"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    },
                    "403": {
                        "description": "Collaborator Already Created"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateCollaboratorById": {
            "post": {
                "summary": "update Colaborator",
                "description": " Update Colaborator Note ",
                "operationId": "d9e84b722ef4ced09b8904919c05232e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "note_id",
                                    "updated_title",
                                    "updated_description"
                                ],
                                "properties": {
                                    "note_id": {
                                        "type": "integer"
                                    },
                                    "updated_title": {
                                        "type": "string"
                                    },
                                    "updated_description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collaborator updated Sucessfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    },
                    "409": {
                        "description": "Collaborator Already Created"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteCollaboratorById": {
            "post": {
                "summary": "delete Colaborator",
                "description": " Delete Colaborator Note ",
                "operationId": "ffd8b9087eb005c206351d65df8e8963",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "note_id",
                                    "email"
                                ],
                                "properties": {
                                    "note_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collaborator deleted Sucessfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    },
                    "404": {
                        "description": "Not a Registered Email"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}